+++ sgl-model-gateway/src/routers/grpc/regular/responses/mod.rs
@@ -1,20 +1,23 @@
 //! Regular gRPC Router `/v1/responses` endpoint implementation
 //!
-//! This module handles all responses-specific logic for the regular (non-Harmony) pipeline including:
-//! - Request validation
-//! - Conversation history and response chain loading
-//! - Streaming support
-//! - MCP tool loop wrapper
-//! - Response persistence
+//! This module handles all responses-specific logic for the regular (non-Harmony) pipeline.
+//!
+//! ## Architecture
+//!
+//! - `handlers` - Entry points: route_responses (thin dispatcher)
+//! - `non_streaming` - Non-streaming execution with MCP tool loop
+//! - `streaming` - Streaming execution with MCP tool loop
+//! - `common` - Shared helpers: ToolLoopState, tool preparation, MCP metadata builders
+//! - `conversions` - Request/response conversion between Responses and Chat formats
+//! - `context` - ResponsesContext and BackgroundTaskInfo
 
-// Module declarations
-pub mod context;
+mod common;
+mod context;
 mod conversions;
 mod handlers;
-pub mod tool_loop;
-pub mod types;
+mod non_streaming;
+mod streaming;
 
 // Public exports
-pub use context::ResponsesContext;
+pub use context::{BackgroundTaskInfo, ResponsesContext};
 pub use handlers::route_responses;
-pub use types::BackgroundTaskInfo;
